% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_console.R
\name{help_console}
\alias{help_console}
\title{Print help file into console}
\usage{
help_console(
  topic,
  package = "replicationOrigins",
  format = c("text", "html", "latex", "Rd"),
  before = NULL,
  after = NULL
)
}
\arguments{
\item{topic}{the command for which help is required}

\item{package}{the package name with the required topic}

\item{format}{output format}

\item{before}{place code before the output e.g. \verb{<blockquote>}}

\item{after}{place code after the output e.g. \verb{</blockquote>}}
}
\value{
formatted function help
}
\description{
From https://stackoverflow.com/questions/12038160/how-to-not-run-an-example-using-roxygen2
based on code by Noam Ross
http://www.noamross.net/archives/2013-06-18-helpconsoleexample/
Stéphane Laurent
https://stackoverflow.com/questions/60468080/
print-an-r-help-file-vignette-as-output-into-an-r-html-notebook
and Michael Sumner (mdsumner)
https://stackoverflow.com/questions/7495685/
how-to-access-the-help-documentation-rd-source-files-in-r
}
