% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericVectorAlign.R
\name{numericVectorAlign}
\alias{numericVectorAlign}
\title{Compares two numeric vectors using Nedleman-Wunsh alignment algorithm}
\usage{
numericVectorAlign(vec1, vec2, gap_penalty = 1e+06, sort = FALSE)
}
\arguments{
\item{vec1}{first numerical vector (e.g. genomic coordinates)}

\item{vec2}{second numerical vector}

\item{gap_penalty}{penalty for gap insertion (default:1000000)}

\item{sort}{if True the vectors will be position sorted before conducting the alignemnt (in this case they cannot contain NA values) (default: False)}
}
\value{
Returns a list of  variables: Score, AlignTab and Cons. Score is the genreral similarity score between the sequences. AlignTab is a table providing paired coordinates from both vectors with NA indicating gaps. Cons is the consensus vector obtained by averaging the paired coordinates.
}
\description{
Compares two numeric vectors using Nedleman-Wunsh alignment algorithm
}
