% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectPeaksOKseq.R
\name{detectPeaksOKseq}
\alias{detectPeaksOKseq}
\title{Detection of RFD profile changes associated with DNA replication origins, based on OK-seq data}
\usage{
detectPeaksOKseq(Data, IntSize = 2e+05, StepSize = 1000)
}
\arguments{
\item{Data}{data frame containing position-specific RFD values. The format is as following, 4 columns chr, start, end, RFD}

\item{IntSize}{size of the window used to calculate linear regression (default:200000)}

\item{StepSize}{distance between intervals used to calculate linear regression (default:1000)}
}
\value{
Table with coordinates of the RFD shifts associated with DNA replication origins
}
\description{
Detection of RFD profile changes associated with DNA replication origins, based on OK-seq data
}
