% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleNumericVectorAlign.R
\name{multipleNumericVectorAlign}
\alias{multipleNumericVectorAlign}
\title{Performs a comparison of multiple numerical vectors (e.g. genomic coordinates) using an algorithm similar to ClustalW}
\usage{
multipleNumericVectorAlign(vector_list, gap_penalty = 1e+06)
}
\arguments{
\item{vector_list}{list of numeric vectors to be compared}

\item{gap_penalty}{penalty for gap insertion (default:1000000)}
}
\value{
Matrix with matched coordinates from all vectors with NA indicating gaps.
}
\description{
Performs a comparison of multiple numerical vectors (e.g. genomic coordinates) using an algorithm similar to ClustalW
}
